Sync flutter/packages main (~781 commits) while keeping fork-only
StoreKit 2 APIs: showManageSubscriptions and willAutoRenew, plus the
InAppPurchaseStoreKitPlatformAddition re-export from in_app_purchase.

Removed redundant isEligibleForIntroOffer fork API in favor of
upstream isIntroductoryOfferEligible. Bumped in_app_purchase_storekit
to 0.4.11 and in_app_purchase to 3.3.1 for fork carry-over changes.

+++ packages/pigeon/platform_tests/test_plugin/android/src/test/kotlin/com/example/test_plugin/PrimitiveTest.kt
@@ -1,4 +1,4 @@
-// Copyright 2013 The Flutter Authors. All rights reserved.
+// Copyright 2013 The Flutter Authors
 // Use of this source code is governed by a BSD-style license that can be
 // found in the LICENSE file.
 
@@ -42,7 +42,7 @@ class PrimitiveTest {
     }
 
     verify { binaryMessenger.setMessageHandler(channelName, handlerSlot.captured) }
-    verify { api.anInt(input.toLong()) }
+    verify { api.anInt(input) }
   }
 
   @Test
